refactor: global/exception 폴터 구조 변경 (#29)
