+++ src/main/java/eightbit/moyeohaeng/global/exception/ErrorCode.java
@@ -1,4 +1,4 @@
-package eightbit.moyeohaeng.global.exception.common;
+package eightbit.moyeohaeng.global.exception;
 
 import org.springframework.http.HttpStatus;
